producer/requirements.txt
pydantic
kafka-python
docker